api/requirements.txt
flask
pyodbc
azure-functions